HyperVM Refactor
