+++ src/Neo.HyperVM/ArrayStackItem.cpp
