+++ src/Neo.HyperVM/ArrayStackItem.h
