+++ src/Neo.HyperVM/BigIntegerBuilder.cpp
