+++ src/Neo.HyperVM/BigIntegerBuilder.h
