+++ src/Neo.HyperVM/BoolStackItem.cpp
