+++ src/Neo.HyperVM/BoolStackItem.h
