+++ src/Neo.HyperVM/ByteArrayStackItem.cpp
