+++ src/Neo.HyperVM/ByteArrayStackItem.h
