+++ src/Neo.HyperVM/EStackItemType.h
