+++ src/Neo.HyperVM/IStackItem.cpp
