+++ src/Neo.HyperVM/IStackItem.h
