+++ src/Neo.HyperVM/IntegerStackItem.cpp
