+++ src/Neo.HyperVM/IntegerStackItem.h
