+++ src/Neo.HyperVM/InteropStackItem.h
