+++ src/Neo.HyperVM/MapStackItem.cpp
