+++ src/Neo.HyperVM/MapStackItem.h
