+++ src/Neo.HyperVM/Stack.cpp
