+++ src/Neo.HyperVM/Stack.h
