+++ src/Neo.HyperVM/StackItems.h
